refactor: config 패키지 구조 변경에 따른 import 경로 및 API 시그니처 수정
- BusinessException·ErrorCode·ApiResponse 이동(config → config.exception/common) 반영
- listRootFolders 시그니처 변경(memberId 2개) 에 맞춰 호출부 수정
